фикс изменений
