+++ src/components/OrdersReady/OrderReady.tsx
@@ -1,20 +1,29 @@
-import React from 'react';
-import OrderReadyCss from './OrderReady.module.css';
+import React from "react";
+import OrderReadyCss from "./OrderReady.module.css";
+import { useSelector, useDispatch } from "react-redux";
+import { useEffect } from "react";
+import { connectFeed, disconnectFeed } from "../../services/actions/socket-action";
+import { FEED_SOCKET_URL_All } from "../../utils/vars";
 
 const OrderReady = (): React.JSX.Element => {
-	return (
-		<div>
-			<span className="text text_type_main-default">Готовы:</span>
-			<ul className={OrderReadyCss.list}>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-				<li className={`${OrderReadyCss.listItem} text text_type_main-default`}>034533</li>
-			</ul>
-		</div>
-	);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(connectFeed(FEED_SOCKET_URL_All));
+    return () => {
+      dispatch(disconnectFeed());
+    };
+  }, [dispatch]);
+
+  const arrayAllOrdersSocket = useSelector((store) => store.feedReducer.orders.orders);
+  const doneOrders = arrayAllOrdersSocket?.filter((item) => item.status === "done").slice(0, 10);
+
+  return (
+    <div>
+      <span className="text text_type_main-default">Готовы:</span>
+      <ul className={OrderReadyCss.list}>{doneOrders?.length > 0 && doneOrders.map((item) => <li className={`${OrderReadyCss.listItem} text text_type_main-default`}>{item.number}</li>)}</ul>
+    </div>
+  );
 };
 
 export default OrderReady;
